Split managers into a new module (#37)
