+++ onvif/client.py
@@ -1,134 +1,46 @@
 """ONVIF Client."""
 from __future__ import annotations
 
-from abc import abstractmethod
 import asyncio
-from collections.abc import Awaitable
-import contextlib
 import datetime as dt
-from functools import lru_cache, partial
+from functools import lru_cache
 import logging
 import os.path
-import ssl
-from typing import Any, Callable, Dict, Optional, ParamSpec, Tuple, TypeVar
+from typing import Any, Callable
 
 import httpx
-from httpx import AsyncClient, BasicAuth, DigestAuth, TransportError
+from httpx import AsyncClient, BasicAuth, DigestAuth
 from zeep.cache import SqliteCache
-from zeep.client import AsyncClient as BaseZeepAsyncClient, Settings
-from zeep.exceptions import Fault, XMLParseError, XMLSyntaxError
+from zeep.client import AsyncClient as BaseZeepAsyncClient
 import zeep.helpers
-from zeep.loader import parse_xml
 from zeep.proxy import AsyncServiceProxy
-from zeep.transports import AsyncTransport, Transport
+from zeep.transports import AsyncTransport
 from zeep.wsa import WsAddressingPlugin
 from zeep.wsdl import Document
-from zeep.wsdl.bindings.soap import SoapOperation
 from zeep.wsse.username import UsernameToken
 
 from onvif.definition import SERVICES
 from onvif.exceptions import ONVIFAuthError, ONVIFError, ONVIFTimeoutError
 
-from .util import normalize_url, stringify_onvif_error
+from .const import KEEPALIVE_EXPIRY
+from .managers import NotificationManager, PullPointManager
+from .settings import DEFAULT_SETTINGS
+from .transport import ASYNC_TRANSPORT
+from .util import create_no_verify_ssl_context, path_isfile, utcnow
+from .wrappers import retry_connection_error  # noqa: F401
 
 logger = logging.getLogger("onvif")
 logging.basicConfig(level=logging.INFO)
 logging.getLogger("zeep.client").setLevel(logging.CRITICAL)
 
-_DEFAULT_SETTINGS = Settings()
-_DEFAULT_SETTINGS.strict = False
-_DEFAULT_SETTINGS.xml_huge_tree = True
 
 _WSDL_PATH = os.path.join(os.path.dirname(os.path.dirname(__file__)), "wsdl")
-
 _DEFAULT_TIMEOUT = 90
 _PULLPOINT_TIMEOUT = 90
 _CONNECT_TIMEOUT = 30
 _READ_TIMEOUT = 90
 _WRITE_TIMEOUT = 90
-
-_RENEWAL_PERCENTAGE = 0.8
-
-KEEPALIVE_EXPIRY = 4
-BACKOFF_TIME = KEEPALIVE_EXPIRY + 0.5
-HTTPX_LIMITS = httpx.Limits(keepalive_expiry=4)
-
-SUBSCRIPTION_ERRORS = (Fault, asyncio.TimeoutError, TransportError)
-RENEW_ERRORS = (ONVIFError, httpx.RequestError, XMLParseError, *SUBSCRIPTION_ERRORS)
-SUBSCRIPTION_RESTART_INTERVAL_ON_ERROR = dt.timedelta(seconds=40)
-
-DEFAULT_ATTEMPTS = 2
-
-P = ParamSpec("P")
-T = TypeVar("T")
-
-
-def retry_connection_error(
-    attempts: int = DEFAULT_ATTEMPTS,
-) -> Callable[[Callable[P, Awaitable[T]]], Callable[P, Awaitable[T]]]:
-    """Define a wrapper to retry on connection error."""
-
-    def _decorator_retry_connection_error(
-        func: Callable[P, Awaitable[T]]
-    ) -> Callable[P, Awaitable[T]]:
-        """Define a wrapper to retry on connection error.
-
-        The remote server is allowed to disconnect us any time so
-        we need to retry the operation.
-        """
-
-        async def _async_wrap_connection_error_retry(  # type: ignore[return]
-            *args: P.args, **kwargs: P.kwargs
-        ) -> T:
-            for attempt in range(attempts):
-                try:
-                    return await func(*args, **kwargs)
-                except httpx.RequestError as ex:
-                    #
-                    # We should only need to retry on RemoteProtocolError but some cameras
-                    # are flakey and sometimes do not respond to the Renew request so we
-                    # retry on RequestError as well.
-                    #
-                    # For RemoteProtocolError:
-                    # http://datatracker.ietf.org/doc/html/rfc2616#section-8.1.4 allows the server
-                    # to close the connection at any time, we treat this as a normal and try again
-                    # once since we do not want to declare the camera as not supporting PullPoint
-                    # if it just happened to close the connection at the wrong time.
-                    if attempt == attempts - 1:
-                        raise
-                    logger.debug(
-                        "Error: %s while calling %s, backing off: %s, retrying...",
-                        ex,
-                        func,
-                        BACKOFF_TIME,
-                        exc_info=True,
-                    )
-                    await asyncio.sleep(BACKOFF_TIME)
-
-        return _async_wrap_connection_error_retry
-
-    return _decorator_retry_connection_error
-
-
-def create_no_verify_ssl_context() -> ssl.SSLContext:
-    """Return an SSL context that does not verify the server certificate.
-    This is a copy of aiohttp's create_default_context() function, with the
-    ssl verify turned off and old SSL versions enabled.
-
-    https://github.com/aio-libs/aiohttp/blob/33953f110e97eecc707e1402daa8d543f38a189b/aiohttp/connector.py#L911
-    """
-    sslcontext = ssl.SSLContext(ssl.PROTOCOL_TLS_CLIENT)
-    sslcontext.check_hostname = False
-    sslcontext.verify_mode = ssl.CERT_NONE
-    # Allow all ciphers rather than only Python 3.10 default
-    sslcontext.set_ciphers("DEFAULT")
-    with contextlib.suppress(AttributeError):
-        # This only works for OpenSSL >= 1.0.0
-        sslcontext.options |= ssl.OP_NO_COMPRESSION
-    sslcontext.set_default_verify_paths()
-    return sslcontext
-
-
+_HTTPX_LIMITS = httpx.Limits(keepalive_expiry=KEEPALIVE_EXPIRY)
 _NO_VERIFY_SSL_CONTEXT = create_no_verify_ssl_context()
 
 
@@ -168,30 +80,10 @@ class UsernameDigestTokenDtDiff(UsernameToken):
         return result
 
 
-class AsyncSafeTransport(Transport):
-    """A transport that blocks all remote I/O for zeep."""
-
-    def load(self, url: str) -> None:
-        """Load the given XML document."""
-        if not _path_isfile(url):
-            raise RuntimeError(f"Loading {url} is not supported in async mode")
-        # Ideally this would happen in the executor but the library
-        # does not call this from a coroutine so the best we can do
-        # without a major refactor is to cache this so it only happens
-        # once per process at startup. Previously it would happen once
-        # per service per camera per setup which is a lot of blocking
-        # I/O in the event loop so this is a major improvement.
-        with open(os.path.expanduser(url), "rb") as fh:
-            return fh.read()
-
-
-_ASYNC_TRANSPORT = AsyncSafeTransport()
-
-
 @lru_cache(maxsize=128)
 def _cached_document(url: str) -> Document:
     """Load external XML document from disk."""
-    return Document(url, _ASYNC_TRANSPORT, settings=_DEFAULT_SETTINGS)
+    return Document(url, ASYNC_TRANSPORT, settings=DEFAULT_SETTINGS)
 
 
 class ZeepAsyncClient(BaseZeepAsyncClient):
@@ -210,11 +102,6 @@ class ZeepAsyncClient(BaseZeepAsyncClient):
                 "are: %s" % (", ".join(self.wsdl.bindings.keys()))
             )
         return AsyncServiceProxy(self, binding, address=address)
-
-
-# This does blocking I/O (stat) so we cache the result
-# to minimize the impact of the blocking I/O.
-_path_isfile = lru_cache(maxsize=128)(os.path.isfile)
 
 
 class ONVIFService:
@@ -263,7 +150,7 @@ class ONVIFService:
         read_timeout: int | None = None,
         write_timeout: int | None = None,
     ) -> None:
-        if not _path_isfile(url):
+        if not path_isfile(url):
             raise ONVIFError("%s doesn`t exist!" % url)
 
         self.url = url
@@ -284,13 +171,13 @@ class ONVIFService:
             write=write_timeout or _WRITE_TIMEOUT,
         )
         client = AsyncClient(
-            verify=_NO_VERIFY_SSL_CONTEXT, timeout=timeouts, limits=HTTPX_LIMITS
+            verify=_NO_VERIFY_SSL_CONTEXT, timeout=timeouts, limits=_HTTPX_LIMITS
         )
         # The wsdl client should never actually be used, but it is required
         # to avoid creating another ssl context since the underlying code
         # will try to create a new one if it doesn't exist.
         wsdl_client = httpx.Client(
-            verify=_NO_VERIFY_SSL_CONTEXT, timeout=timeouts, limits=HTTPX_LIMITS
+            verify=_NO_VERIFY_SSL_CONTEXT, timeout=timeouts, limits=_HTTPX_LIMITS
         )
         self.transport = (
             AsyncTransport(client=client, wsdl_client=wsdl_client)
@@ -309,7 +196,7 @@ class ONVIFService:
 
     async def setup(self):
         """Setup the transport."""
-        settings = _DEFAULT_SETTINGS
+        settings = DEFAULT_SETTINGS
         binding_name = self.binding_name
         wsse = UsernameDigestTokenDtDiff(
             self.user, self.passwd, dt_diff=self.dt_diff, use_digest=self.encrypt
@@ -387,302 +274,6 @@ class ONVIFService:
         if name.startswith("authless_"):
             return service_wrapper(getattr(self.ws_client_authless, name.split("_")[1]))
         return service_wrapper(getattr(self.ws_client, name))
-
-
-class BaseManager:
-    """Base class for notification and pull point managers."""
-
-    def __init__(
-        self,
-        device: ONVIFCamera,
-        interval: dt.timedelta,
-        subscription_lost_callback: Callable[[], None],
-    ) -> None:
-        """Initialize the notification processor."""
-        self._operation: SoapOperation | None = None
-        self._device = device
-        self._interval = interval
-        self._renew_lock = asyncio.Lock()
-        self._subscription: ONVIFService | None = None
-        self._restart_or_renew_task: asyncio.Task | None = None
-        self._loop = asyncio.get_event_loop()
-        self._shutdown = False
-        self._subscription_lost_callback = subscription_lost_callback
-        self._cancel_subscription_renew: asyncio.TimerHandle | None = None
-
-    @property
-    def closed(self) -> bool:
-        """Return True if the manager is closed."""
-        return not self._subscription or self._subscription.transport.client.is_closed
-
-    async def start(self) -> None:
-        """Setup the manager."""
-        renewal_call_at = await self._start()
-        self._schedule_subscription_renew(renewal_call_at)
-        return self._subscription
-
-    def pause(self) -> None:
-        """Pause the manager."""
-        self._cancel_renewals()
-
-    def resume(self) -> None:
-        """Resume the manager."""
-        self._schedule_subscription_renew(self._loop.time())
-
-    async def stop(self) -> None:
-        """Stop the manager."""
-        logger.debug("%s: Stop the notification manager", self._device.host)
-        self._cancel_renewals()
-        assert self._subscription, "Call start first"
-        await self._subscription.Unsubscribe()
-
-    async def shutdown(self) -> None:
-        """Shutdown the manager.
-
-        This method is irreversible.
-        """
-        self._shutdown = True
-        if self._restart_or_renew_task:
-            self._restart_or_renew_task.cancel()
-        logger.debug("%s: Shutdown the notification manager", self._device.host)
-        await self.stop()
-
-    @abstractmethod
-    async def _start(self) -> float:
-        """Setup the processor. Returns the next renewal call at time."""
-
-    async def _set_synchronization_point(self, service: ONVIFService) -> float:
-        """Set the synchronization point."""
-        try:
-            await service.SetSynchronizationPoint()
-        except (Fault, asyncio.TimeoutError, TransportError, TypeError):
-            logger.debug("%s: SetSynchronizationPoint failed", self._service.url)
-
-    def _cancel_renewals(self) -> None:
-        """Cancel any pending renewals."""
-        if self._cancel_subscription_renew:
-            self._cancel_subscription_renew.cancel()
-            self._cancel_subscription_renew = None
-
-    def _calculate_next_renewal_call_at(self, result: Any | None) -> float:
-        """Calculate the next renewal call_at."""
-        current_time: dt.datetime | None = result.CurrentTime
-        termination_time: dt.datetime | None = result.TerminationTime
-        if termination_time and current_time:
-            delay = termination_time - current_time
-        else:
-            delay = self._interval
-        delay_seconds = delay.total_seconds() * _RENEWAL_PERCENTAGE
-        logger.debug(
-            "%s: Renew notification subscription in %s seconds",
-            self._device.host,
-            delay_seconds,
-        )
-        return self._loop.time() + delay_seconds
-
-    def _schedule_subscription_renew(self, when: float) -> None:
-        """Schedule notify subscription renewal."""
-        self._cancel_renewals()
-        self._cancel_subscription_renew = self._loop.call_at(
-            when,
-            self._run_restart_or_renew,
-        )
-
-    def _run_restart_or_renew(self) -> None:
-        """Create a background task."""
-        if self._restart_or_renew_task and not self._restart_or_renew_task.done():
-            logger.debug("%s: Notify renew already in progress", self._device.host)
-            return
-        self._restart_or_renew_task = asyncio.create_task(
-            self._renew_or_restart_subscription()
-        )
-
-    async def _restart_subscription(self) -> bool:
-        """Restart the notify subscription assuming the camera rebooted."""
-        self._cancel_renewals()
-        return await self._start()
-
-    @retry_connection_error()
-    async def _call_subscription_renew(self) -> float:
-        """Call notify subscription Renew."""
-        device = self._device
-        logger.debug("%s: Renew the notification manager", device.host)
-        return self._calculate_next_renewal_call_at(
-            await self._subscription.Renew(
-                device.get_next_termination_time(self._interval)
-            )
-        )
-
-    async def _renew_subscription(self) -> float | None:
-        """Renew notify subscription."""
-        if self.closed or self._shutdown:
-            return None
-        try:
-            return await self._call_subscription_renew()
-        except RENEW_ERRORS as err:
-            self._subscription_lost_callback()
-            logger.debug(
-                "%s: Failed to renew notify subscription %s",
-                self._device.host,
-                stringify_onvif_error(err),
-            )
-        return None
-
-    async def _renew_or_restart_subscription(self) -> None:
-        """Renew or start notify subscription."""
-        if self._shutdown:
-            return
-        renewal_call_at = None
-        try:
-            renewal_call_at = (
-                await self._renew_subscription() or await self._restart_subscription()
-            )
-        finally:
-            self._schedule_subscription_renew(
-                renewal_call_at
-                or self._loop.time()
-                + SUBSCRIPTION_RESTART_INTERVAL_ON_ERROR.total_seconds()
-            )
-
-
-class NotificationManager(BaseManager):
-    """Manager to process notifications."""
-
-    def __init__(
-        self,
-        device: ONVIFCamera,
-        address: str,
-        interval: dt.timedelta,
-        subscription_lost_callback: Callable[[], None],
-    ) -> None:
-        """Initialize the notification processor."""
-        self._address = address
-        super().__init__(device, interval, subscription_lost_callback)
-
-    async def _start(self) -> float:
-        """Start the notification processor.
-
-        Returns the next renewal call at time.
-        """
-        device = self._device
-        logger.debug("%s: Setup the notification manager", device.host)
-        notify_service = await device.create_notification_service()
-        time_str = device.get_next_termination_time(self._interval)
-        result = await notify_service.Subscribe(
-            {
-                "InitialTerminationTime": time_str,
-                "ConsumerReference": {"Address": self._address},
-            }
-        )
-        # pylint: disable=protected-access
-        device.xaddrs[
-            "http://www.onvif.org/ver10/events/wsdl/NotificationConsumer"
-        ] = normalize_url(result.SubscriptionReference.Address._value_1)
-        # Create subscription manager
-        # 5.2.3 BASIC NOTIFICATION INTERFACE - NOTIFY
-        # Call SetSynchronizationPoint to generate a notification message
-        # to ensure the webhooks are working.
-        #
-        # If this fails this is OK as it just means we will switch
-        # to webhook later when the first notification is received.
-        service = await self._device.create_onvif_service(
-            "pullpoint", port_type="NotificationConsumer"
-        )
-        self._operation = service.document.bindings[service.binding_name].get(
-            "PullMessages"
-        )
-        self._subscription = await device.create_subscription_service(
-            "NotificationConsumer"
-        )
-        if device.has_broken_relative_time(
-            self._interval,
-            result.CurrentTime,
-            result.TerminationTime,
-        ):
-            # If we determine the device has broken relative timestamps, we switch
-            # to using absolute timestamps and renew the subscription.
-            result = await self._subscription.Renew(
-                device.get_next_termination_time(self._interval)
-            )
-        renewal_call_at = self._calculate_next_renewal_call_at(result)
-        logger.debug("%s: Start the notification manager", self._device.host)
-        await self._set_synchronization_point(service)
-        return renewal_call_at
-
-    def process(self, content: bytes) -> Any | None:
-        """Process a notification message."""
-        if not self._operation:
-            logger.debug("%s: Notifications not setup", self._device.host)
-            return
-        try:
-            envelope = parse_xml(
-                content,  # type: ignore[arg-type]
-                _ASYNC_TRANSPORT,
-                settings=_DEFAULT_SETTINGS,
-            )
-        except XMLSyntaxError as exc:
-            logger.error("Received invalid XML: %s", exc)
-            return None
-        return self._operation.process_reply(envelope)
-
-
-class PullPointManager(BaseManager):
-    """Manager for PullPoint."""
-
-    def __init__(
-        self,
-        device: ONVIFCamera,
-        interval: dt.timedelta,
-        subscription_lost_callback: Callable[[], None],
-    ) -> None:
-        """Initialize the PullPoint processor."""
-        super().__init__(device, interval, subscription_lost_callback)
-        self._service: ONVIFService | None = None
-
-    async def _start(self) -> float:
-        """Start the PullPoint manager.
-
-        Returns the next renewal call at time.
-        """
-        device = self._device
-        logger.debug("%s: Setup the PullPoint manager", device.host)
-        events_service = await device.create_events_service()
-        result = await events_service.CreatePullPointSubscription(
-            {
-                "InitialTerminationTime": device.get_next_termination_time(
-                    self._interval
-                ),
-            }
-        )
-        # pylint: disable=protected-access
-        device.xaddrs[
-            "http://www.onvif.org/ver10/events/wsdl/PullPointSubscription"
-        ] = normalize_url(result.SubscriptionReference.Address._value_1)
-        # Create subscription manager
-        self._subscription = await device.create_subscription_service(
-            "PullPointSubscription"
-        )
-        # Create the service that will be used to pull messages from the device.
-        self._service = await device.create_pullpoint_service()
-        if device.has_broken_relative_time(
-            self._interval, result.CurrentTime, result.TerminationTime
-        ):
-            # If we determine the device has broken relative timestamps, we switch
-            # to using absolute timestamps and renew the subscription.
-            result = await self._subscription.Renew(
-                device.get_next_termination_time(self._interval)
-            )
-        renewal_call_at = self._calculate_next_renewal_call_at(result)
-        logger.debug("%s: Start the notification manager", self._device.host)
-        await self._set_synchronization_point(self._service)
-        return renewal_call_at
-
-    def get_service(self) -> ONVIFService:
-        """Return the pullpoint service."""
-        return self._service
-
-
-_utcnow: partial[dt.datetime] = partial(dt.datetime.now, dt.timezone.utc)
 
 
 class ONVIFCamera:
@@ -840,7 +431,7 @@ class ONVIFCamera:
         """Calculate subscription absolute termination time."""
         if not self._has_broken_relative_timestamps:
             return f"PT{int(duration.total_seconds())}S"
-        absolute_time: dt.datetime = _utcnow() + duration
+        absolute_time: dt.datetime = utcnow() + duration
         if dt_diff := self.dt_diff:
             absolute_time += dt_diff
         return absolute_time.isoformat(timespec="seconds").replace("+00:00", "Z")
@@ -932,7 +523,7 @@ class ONVIFCamera:
             namespace += "/" + port_type
 
         wsdlpath = os.path.join(self.wsdl_dir, wsdl_file)
-        if not _path_isfile(wsdlpath):
+        if not path_isfile(wsdlpath):
             raise ONVIFError("No such file: %s" % wsdlpath)
 
         # XAddr for devicemgmt is fixed:
