+++ onvif/const.py
@@ -1,0 +1,5 @@
+"""ONVIF constants."""
+
+DEFAULT_ATTEMPTS = 2
+KEEPALIVE_EXPIRY = 4
+BACKOFF_TIME = KEEPALIVE_EXPIRY + 0.5
